Only save user features for needed users
The evaluation code is slow, but not because of computations.
I did the profiling and found that the most time is spent in
searching pandas dataframes for user features and rewards.
As a first step - I reduced the number of stored data because
we stored also features of users not in the experiment. This already
gave a large speedup.